fixed parser & implemented hard version of hw6 & added readme to hw6

+++ HW6/TripleExpression/Add.java
@@ -5,7 +5,7 @@ public class Add extends BinaryOperator implements TripleExpression {
         super(first, second);
     }
 
-    protected double apply(double a, double b) {
+    protected int apply(int a, int b) {
         return a + b;
     }
 }
