+++ HW6/TripleExpression/Subtract.java
@@ -5,7 +5,7 @@ public class Subtract extends BinaryOperator implements TripleExpression {
         super(first, second);
     }
 
-    protected double apply(double a, double b) {
+    protected int apply(int a, int b) {
         return a - b;
     }
 }
