+++ HW6/TripleExpression/TripleExpression.java
@@ -1,5 +1,5 @@
 package TripleExpression;
 
 public interface TripleExpression {
-    double evaluate(double x, double y, double z);
+    int evaluate(int x, int y, int z);
 }
