+++ HW6/TripleExpression/Variable.java
@@ -6,7 +6,7 @@ public class Variable implements TripleExpression {
         name = nam;
     }
 
-    public double evaluate(double x, double y, double z) {
+    public int evaluate(int x, int y, int z) {
         if (name.equals("x")) {
             return x;
         } else if (name.equals("y")) {
